reorganize routes into two files and reflect changes in server.js
