--- controllers/AlbumController.js
@@ -1,56 +1,0 @@
-var express = require('express');
-var router = express.Router();
-var models = require('../models');
-var auth = require('../lib/helpers');
-
-// searches for all albums that match the search parameters
-router.post('/search', auth.validate, function(req, res) {
-	var whereObj;
-	if (req.body.type === 'title') {
-		whereObj = {
-			$or: [
-				{ title : { $eq: req.body.query } },
-				{ title : { like: req.body.query + ' %' } },
-				{ title: { like: '% ' + req.body.query } },
-				{ title: { like: '% ' + req.body.query + ' %' } }
-			]
-		};
-	} else if (req.body.type === 'artist') {
-		whereObj = {
-			$or: [
-				{ '$Artist.artist_name$' : { $eq: req.body.query } },
-				{ '$Artist.artist_name$' : { like: req.body.query + ' %' } },
-				{ '$Artist.artist_name$': { like: '% ' + req.body.query } },
-				{ '$Artist.artist_name$': { like: '% ' + req.body.query + ' %' } }
-			]
-		};
-	}
-
-	models.Album.findAll({
-		where: whereObj,
-		limit: 100,
-		order: [
-			['createdAt', 'DESC']
-		],
-		include: [{
-			model: models.Artist,
-			required: true
-		}, {
-			model: models.Genre,
-			required: true
-		}, {
-			model: models.Label,
-			required: true	
-		}]
-	}).then(function(albumData) {
-		var albumObj = {
-			albums: albumData,
-		};
-		//res.render('album', albumObj);
-		res.json(albumObj);
-	}).catch(function(err) {
-		res.json(err);
-	});
-});
-
-module.exports = router;
