+++ controllers/ApiController.js
@@ -5,9 +5,10 @@ var Sequelize = require('sequelize');
 var auth = require('../lib/helpers');
 var jwt = require('jsonwebtoken');
 var bcrypt = require('bcryptjs');
+var querystring = require('querystring');
 
 // creates a new user
-router.post('/register', function(req, res) {
+router.post('/user/register', function(req, res) {
 	var hashedPassword = bcrypt.hashSync(req.body.password, 8);
 	models.User.create({ 
 		username: req.body.username,
@@ -23,29 +24,8 @@ router.post('/register', function(req, res) {
 	});
 });
 
-router.post('/login', function(req, res) {
-	models.User.findOne({ 
-		where: {
-			$or: [
-				{ email : { $eq: req.body.login } },
-				{ username : { $eq: req.body.login } },
-			]
-		}
-	}).then(function (user) {
-		// if no user is returned, then... d'oh
-		if (!user.dataValues) auth.notFound();
-		auth.handler(req, res, 'login');
-	}).catch(function(err) {
-		return res.status(500).send('Error on the server.');
-	});
-});
-
-router.get('/logout', function(req, res) {
-	res.status(200).send({ auth: false, token: null });
-});
-
 // updates a user's information
-router.put('/update/:id', auth.validate, function(req, res) {
+router.put('/user/:id', auth.validate, function(req, res) {
 	if (req.userId !== req.params.id)
 		return res.status(401).send('You aren\'t authorized to do this!');
 	models.User.update({ 
@@ -65,7 +45,7 @@ router.put('/update/:id', auth.validate, function(req, res) {
 });
 
 // adds an album to the current user's collection
-router.post('/:userId/add/:albumId', auth.validate, function(req, res) {
+router.post('/user/:userId/:albumId', auth.validate, function(req, res) {
 	if(req.userId !== req.params.userId)
 		res.redirect(401, '/login');
 	models.UserAlbum.create({
@@ -79,7 +59,7 @@ router.post('/:userId/add/:albumId', auth.validate, function(req, res) {
 });
 
 // searches for albums in the current user's collection
-router.post('/:id/search', auth.validate, function(req, res) {	
+router.post('/user/:id/search', auth.validate, function(req, res) {	
 	var whereObj;
 	if (req.body.type === 'title') {
 		whereObj = {
@@ -129,6 +109,106 @@ router.post('/:id/search', auth.validate, function(req, res) {
 		};
 		//res.render('user', userObj);
 		res.json(userObj);
+	}).catch(function(err) {
+		res.json(err);
+	});
+});
+
+// creates a new post for the current user
+router.post('/post/:userId/:albumId', auth.validate, function(req, res) {
+	if (req.userId !== req.params.userId) {
+		return res.redirect(401, '/login');
+	}
+	models.Post.create({ 
+		body: req.body.body,
+		isPublic: req.body.isPublic,
+		UserId: req.params.userId,
+		AlbumId: req.params.albumId
+	}).then(function(post) {
+		res.redirect(200, '/user/' + req.username);
+	}).catch(function(err) {
+		res.json(err);
+	});
+});
+
+// updates a post for the current user
+router.put('/post/:postId/:userId', auth.validate, function(req, res) {
+	if (req.userId !== req.params.userId)
+		return res.redirect(401, '/login');
+	models.Post.update({ 
+		body: req.body.body,
+		isPublic: req.body.isPublic,
+	}, {
+		where: {
+			id: req.params.id
+		}
+	}).then(function(post) {
+		res.json(post);
+	}).catch(function(err) {
+		res.json(err);
+	});
+});
+
+// deletes a post
+router.delete('/post/:userId/:postId', auth.validate, function(req, res) {
+	if (req.userId !== req.params.userId)
+		return res.redirect(401, '/login');
+	models.Post.destroy({
+		where: {
+			id: req.params.postId,
+		}
+	}).then(function(result) {
+		res.json(result);
+	}).catch(function(err) {
+		res.json(err);
+	});
+});
+
+// searches for all albums that match the search parameters
+router.post('/album/search', auth.validate, function(req, res) {
+	var whereObj;
+	if (req.body.type === 'title') {
+		whereObj = {
+			$or: [
+				{ title : { $eq: req.body.query } },
+				{ title : { like: req.body.query + ' %' } },
+				{ title: { like: '% ' + req.body.query } },
+				{ title: { like: '% ' + req.body.query + ' %' } }
+			]
+		};
+	} else if (req.body.type === 'artist') {
+		whereObj = {
+			$or: [
+				{ '$Artist.artist_name$' : { $eq: req.body.query } },
+				{ '$Artist.artist_name$' : { like: req.body.query + ' %' } },
+				{ '$Artist.artist_name$': { like: '% ' + req.body.query } },
+				{ '$Artist.artist_name$': { like: '% ' + req.body.query + ' %' } }
+			]
+		};
+	}
+
+	models.Album.findAll({
+		where: whereObj,
+		limit: 100,
+		order: [
+			['createdAt', 'DESC']
+		],
+		include: [{
+			model: models.Artist,
+			required: true
+		}, {
+			model: models.Genre,
+			required: true
+		}, {
+			model: models.Label,
+			required: true	
+		}]
+	}).then(function(albumData) {
+		var albumObj = {
+			albums: albumData,
+		};
+		//res.render('search', albumObj);
+		res.json(albumObj);
 	}).catch(function(err) {
 		res.json(err);
 	});
